Created a debug interface. Fixed bugs. TODO: fix simulation step on GPU

+++ MpiController/src/MPI_Node.hpp
@@ -37,7 +37,7 @@ protected:
     void setComputationalModelEnv(ComputationalModel::NODE_TYPE node_type);
     void setLocalMPI_ids(const size_t globalId, size_t& localIdx, size_t& localIdy);
     void setLocalMPI_ids(const int globalId, int& localIdx, int& localIdy);
-    size_t getGlobalMPIid(size_t mpi_id_x, size_t mpi_id_y);
+    int getGlobalMPIid(int mpi_id_x, int mpi_id_y);
     void finalBarrierSync();
 
 protected:
